vet_rag_project/requirements.txt
faiss-cpu
sentence-transformers
openai
gradio
numpy
torch
